+++ src/utils/DownloardQueue.ts
@@ -19,7 +19,7 @@ export class DownloardQueue {
     }
 
     private Dealer(cb: (file: string) => void, word: string, dfile: string, progress: number, state: string, why?: string) {
-        console.log(`${(progress * 100).toFixed(2)}% of ${dfile} was ${state} to download!`);
+        console.log(`${(progress * 100).toFixed(2)}% of ${dfile} was ${state} to download`);
         let gApp = globalVar.app;
 		let ext = path.extname(dfile);
 		console.log(ext);